SDPF/requirements.txt
Flask==3.1.0
h5py==3.10.0
matplotlib==3.7.2
numpy==1.25.2
pymongo==4.10.1
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
ruamel.base==1.0.0
scikit_learn==1.3.0
scipy==1.14.1
setuptools==68.0.0
